// Teleopti.Ccc.Web/Teleopti.Ccc.Web/WFM/src/app/intraday/types/intraday-traffic-data.ts
// Generated by https://quicktype.io

export interface IntradayTrafficData {
	FirstIntervalStart: string;
	FirstIntervalEnd: string;
	LatestActualIntervalStart: string;
	LatestActualIntervalEnd: string;
	Summary: IntradayTrafficSummaryData;
	DataSeries: IntradayTrafficDataSeries;
	IncomingTrafficHasData: boolean;
}

export interface IntradayTrafficDataSeries {
	AverageSpeedOfAnswer: any[];
	Time: any[];
	ForecastedCalls: any[];
	ForecastedAverageHandleTime: any[];
	AverageHandleTime: any[];
	CalculatedCalls: any[];
	AbandonedRate: any[];
	ServiceLevel: any[];
}
export interface IntradayTrafficSummaryItem {
	Heading: string;
	Forecasted: number;
	Actual: number;
	Difference: number;
}

export interface IntradayTrafficSummaryData {
	ForecastedCalls: number;
	ForecastedAverageHandleTime: number;
	ForecastedHandleTime: number;
	CalculatedCalls: number;
	AverageHandleTime: number;
	HandleTime: number;
	ForecastedActualCallsDiff: number;
	ForecastedActualHandleTimeDiff: number;
	AverageSpeedOfAnswer: number;
	SpeedOfAnswer: number;
	AnsweredCalls: number;
	ServiceLevel: number;
	AnsweredCallsWithinSL: number;
	AbandonRate: number;
	AbandonedCalls: number;
}